more general interface for actions

+++ cpp/unreal_plugins/SimulationController/Source/SimulationController/Agent.h
@@ -29,7 +29,7 @@ public:
     virtual std::map<std::string, Box> getObservationSpace() const = 0;
     virtual std::map<std::string, Box> getStepInfoSpace() const = 0;   
  
-    virtual void applyAction(const std::map<std::string, std::vector<float>>& action) = 0;
+    virtual void applyAction(const std::map<std::string, std::vector<uint8_t>>& action) = 0;
     virtual std::map<std::string, std::vector<uint8_t>> getObservation() const = 0;
     virtual std::map<std::string, std::vector<uint8_t>> getStepInfo() const = 0;
