+++ cpp/unreal_plugins/SimulationController/Source/SimulationController/CameraAgent.h
@@ -33,7 +33,7 @@ public:
     std::map<std::string, Box> getObservationSpace() const override;
     std::map<std::string, Box> getStepInfoSpace() const override;
 
-    void applyAction(const std::map<std::string, std::vector<float>>& action) override;
+    void applyAction(const std::map<std::string, std::vector<uint8_t>>& action) override;
     std::map<std::string, std::vector<uint8_t>> getObservation() const override;
     std::map<std::string, std::vector<uint8_t>> getStepInfo() const override;
 
@@ -50,5 +50,5 @@ private:
 
     std::unique_ptr<CameraSensor> camera_sensor_;
 
-    std::map<std::string, std::vector<float>> action_;
+    std::map<std::string, std::vector<uint8_t>> action_;
 };
